ZKSPRING-7 Bug in handling of WrongValueException Problem Description: WrongValueException causes doExceptionTranslationFiltering, throw new RuntimeException() , and in org.zkoss.zk.ui.util.Configuration.newEventThreadCleanups(), log.error("Failed to invoke "+klass, t) log it, hence the exception stack trace shows up.
Root Cause:
When UiEngine processing an event, if an exception is thrown, it let some listeners to clear it or further-process by config.newEventThreadCleanups(comp, event, errs, false).

ZkExceptionTranslationListener uses this mechanism to rridge AuthenticationException & AccessDeniedException thrown inside ZK event to popup a login window.

Solution:
Because WrongValueException will be properly handled by UiEngine, we should not throw a new run-time exception to cause that Configuration.newEventThreadCleanups() log this exception.